+++ src/neml2/models/porous_flow/EffectiveSaturation.cxx
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include "neml2/models/reactive_infiltration/EffectiveSaturation.h"
+#include "neml2/models/porous_flow/EffectiveSaturation.h"
 #include "neml2/tensors/functions/clamp.h"
 #include "neml2/tensors/functions/where.h"
