+++ src/neml2/models/porous_flow/PorousFlowCapillaryPressure.cxx
@@ -22,38 +22,29 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include "neml2/models/reactive_infiltration/PorosityPermeabilityRelation.h"
+#include "neml2/models/porous_flow/PorousFlowCapillaryPressure.h"
 
 namespace neml2
 {
 OptionSet
-PorosityPermeabilityRelation::expected_options()
+PorousFlowCapillaryPressure::expected_options()
 {
   OptionSet options = Model::expected_options();
-  options.doc() = "Relate the flow permeability to the porosity, with reference porosity \\f$ "
-                  "\\varphi_o \\f$ and reference permeability \\f$K_o\\f$";
+  options.doc() = "Relate the porous flow capillary pressure to the effective saturation";
 
-  options.set_parameter<TensorName<Scalar>>("reference_permeability") = {TensorName<Scalar>("1")};
-  options.set("reference_permeability").doc() = "the reference permeability";
+  options.set_input("effective_saturation") = VariableName(STATE, "effective_saturation");
+  options.set("effective_saturation").doc() = "The effective saturation";
 
-  options.set_parameter<TensorName<Scalar>>("reference_porosity");
-  options.set("reference_porosity").doc() = "the reference porosity";
-
-  options.set_input("porosity") = VariableName(STATE, "porosity");
-  options.set("porosity").doc() = "porosity";
-
-  options.set_output("permeability") = VariableName(STATE, "permeability");
-  options.set("permeability").doc() = "Porous flow permeability";
+  options.set_output("capillary_pressure") = VariableName(STATE, "capillary_pressure");
+  options.set("capillary_pressure").doc() = "Porous flow capillary pressure.";
 
   return options;
 }
 
-PorosityPermeabilityRelation::PorosityPermeabilityRelation(const OptionSet & options)
+PorousFlowCapillaryPressure::PorousFlowCapillaryPressure(const OptionSet & options)
   : Model(options),
-    _Ko(declare_parameter<Scalar>("Ko", "reference_permeability")),
-    _phio(declare_parameter<Scalar>("phio", "reference_porosity")),
-    _phi(declare_input_variable<Scalar>("porosity")),
-    _K(declare_output_variable<Scalar>("permeability"))
+    _S(declare_input_variable<Scalar>("effective_saturation")),
+    _Pc(declare_output_variable<Scalar>("capillary_pressure"))
 {
 }
